Improve styling and animation when loading custom contribution URL
